+++ src/test/java/org/zalando/nakadi/domain/BatchItemTest.java
@@ -22,8 +22,8 @@ public class BatchItemTest {
     }
 
     @Test
-    public void testBatchItemSizeWithMultByteChar() {
-        final BatchItem item = new BatchItem("{ \"name\": \"香港\"} ",
+    public void testBatchItemSizeWithMultiByteChar() {
+        final BatchItem item = new BatchItem("{ \"name\": \"香港\"} ", true,
                 BatchItem.EmptyInjectionConfiguration.build(1, false),
                 new BatchItem.InjectionConfiguration[BatchItem.Injection.values().length],
                 Collections.emptyList());